+++ plugins/source/aws/resources/services/rds/clusters.go
@@ -31,5 +31,8 @@ func Clusters() *schema.Table {
 				Resolver: resolveRdsClusterTags,
 			},
 		},
+		Relations: []*schema.Table{
+			clusterBacktracks(),
+		},
 	}
 }