+++ plugins/source/azure/resources/services/sql/server_security_alert_policies.go
@@ -1,0 +1,69 @@
+package sql
+
+import (
+	"context"
+	"encoding/json"
+	"net/http"
+
+	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/sql/armsql"
+	"github.com/cloudquery/cloudquery/plugins/source/azure/client"
+	"github.com/cloudquery/plugin-sdk/faker"
+	"github.com/cloudquery/plugin-sdk/schema"
+	"github.com/cloudquery/plugin-sdk/transformers"
+	"github.com/gorilla/mux"
+)
+
+func serverSecurityAlertPolicies() *schema.Table {
+	return &schema.Table{
+		Name:        "azure_sql_server_security_alert_policies",
+		Resolver:    fetchServerSecurityAlertPolicies,
+		Description: "https://learn.microsoft.com/en-us/rest/api/sql/2021-11-01/server-security-alert-policies/list-by-server?tabs=HTTP#serversecurityalertpolicy",
+		Transform:   transformers.TransformWithStruct(&armsql.ServerSecurityAlertPolicy{}, transformers.WithPrimaryKeys("ID")),
+		Columns:     schema.ColumnList{client.SubscriptionID},
+	}
+}
+
+func fetchServerSecurityAlertPolicies(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
+	p := parent.Item.(*armsql.Server)
+	cl := meta.(*client.Client)
+	svc, err := armsql.NewServerSecurityAlertPoliciesClient(cl.SubscriptionId, cl.Creds, cl.Options)
+	if err != nil {
+		return err
+	}
+	group, err := client.ParseResourceGroup(*p.ID)
+	if err != nil {
+		return err
+	}
+	pager := svc.NewListByServerPager(group, *p.Name, nil)
+	for pager.More() {
+		p, err := pager.NextPage(ctx)
+		if err != nil {
+			return err
+		}
+		res <- p.Value
+	}
+	return nil
+}
+
+func createMockServerSecurityAlertPolicies(router *mux.Router) error {
+	var item armsql.ServerSecurityAlertPoliciesClientListByServerResponse
+	if err := faker.FakeObject(&item); err != nil {
+		return err
+	}
+
+	emptyStr := ""
+	item.NextLink = &emptyStr
+
+	router.HandleFunc("/subscriptions/{subscriptionId}/resourceGroups/debug/providers/Microsoft.Sql/servers/test string/securityAlertPolicies", func(w http.ResponseWriter, r *http.Request) {
+		b, err := json.Marshal(&item)
+		if err != nil {
+			http.Error(w, "unable to marshal request: "+err.Error(), http.StatusBadRequest)
+			return
+		}
+		if _, err := w.Write(b); err != nil {
+			http.Error(w, "failed to write", http.StatusBadRequest)
+			return
+		}
+	})
+	return nil
+}